Move GTFSIterator to gtfs.rs
Add some unit tests, fixes a subtle line number count
